Removed config flag for switching between receipt store impls
