+++ rskj-core/src/main/resources/expected.conf
@@ -126,9 +126,6 @@ database = {
         trusted-keys = [<trusted-key>]
         enabled = <bool>
     }
-    receipts = {
-        dataFormatVersion = <version>
-    }
     reset = <bool>
 }
 hello.phrase = <phrase>
