+++ rskj-core/src/main/resources/reference.conf
@@ -161,13 +161,6 @@ database {
         #url = ""
         trusted-keys = []
     }
-
-    receipts {
-        # version of data format of stored receipts
-        # by default, it's "v1"
-        # possible values: ["v1", "v2"]
-        dataFormatVersion: v2
-    }
 }
 
 # Interface to bind peer discovery and wire protocol
