+++ rskj-core/src/test/java/co/rsk/RskContextTest.java
@@ -196,7 +196,6 @@ public class RskContextTest {
 
         doReturn(1).when(testProperties).getNumOfAccountSlots();
         doReturn(true).when(testProperties).fastBlockPropagation();
-        doReturn("").when(testProperties).receiptsDataFormatVersion();
 
         ActivationConfig config = mock(ActivationConfig.class);
         doReturn(config).when(testProperties).getActivationConfig();
